Fixing csv tests
