+++ src/impl/tests/test_scope_mappings.py
@@ -184,36 +184,3 @@ async def test_visible_with_mapping(populated_db):
     assert len(trs) == 1
 
 
-# This is in the Wrong Place, but hey
-@pytest.mark.asyncio
-async def test_csv_generation(populated_db):
-    params = QueryParameters(
-        category_id="residents",
-        indicator_id="residents.residents",
-        srid=3,
-        trid=2,
-        start_date="2020-02-01",
-        duration=3,
-        mdids_only=False,
-    )
-    await run_csv_query(
-        params, pool=populated_db, token_model=TokenModel(sub="TEST",
-                                                          permissions=["admin"])
-    )
-    
-@pytest.mark.asyncio
-async def test_flow_csv_generation(populated_db):
-
-    params = QueryParameters(
-        category_id="relocations",
-        indicator_id="relocations.relocations",
-        srid=3,
-        trid=2,
-        start_date="2020-02-01",
-        duration=3,
-        mdids_only=False,
-    )
-    await run_csv_query(
-        params, pool=populated_db, token_model=TokenModel(sub="TEST",
-                                                          permissions=["admin"])
-    )
